pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "obsidian2confluence"
version = "0.1.0"
description = "One-way synchronization from an Obsidian vault to Confluence Cloud."
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [{ name = "obsidian2confluence maintainers" }]
dependencies = [
  "APScheduler>=3.10.4,<4.0.0",
  "markdown-it-py>=3.0.0,<4.0.0",
  "pydantic>=2.7.4,<3.0.0",
  "python-dotenv>=1.0.1,<2.0.0",
  "python-frontmatter>=1.1.0,<2.0.0",
  "requests>=2.32.3,<3.0.0",
  "tenacity>=8.4.1,<9.0.0",
]

[project.optional-dependencies]
dev = [
  "black>=24.4.2,<25.0.0",
  "ruff>=0.4.9,<0.5.0",
  "pytest>=8.2.2,<9.0.0",
]

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.ruff]
line-length = 100
src = ["app", "tests"]
target-version = "py312"
select = ["E", "F", "I", "UP", "B"]
ignore = ["B008"]  # allow function calls in default args for dependency injection

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra"
testpaths = ["tests"]
